Crypto_currency_scrapping/requirements.txt
pyperclip~=1.8.2
selenium~=4.21.0
Django~=5.0.6
celery~=5.4.0
djangorestframework~=3.15.1